requirements.txt
flask>=0.9
flask-login
flask-openid
flask-mail>=0.7.6
sqlalchemy>=0.7.9
flask-sqlalchemy>=0.16
sqlalchemy-migrate>=0.7.2
flask-whooshalchemy>=0.55a
flask-wtf>=0.8.4
pytz>=2013b
flask-babel>=0.8
flup
